requirements.txt
torch==2.4.0
torchaudio==2.4.0
speechbrain==1.0.2
huggingface_hub==0.23.0
python-dotenv==1.0.1
soundfile==0.12.1
numpy==1.26.4
datasets==2.19.0
librosa==0.10.2